requirements.txt
Flask
srt
google-generativeai
python-dotenv